requirements.txt
streamlit
imageio
multiprocessing-resource_sharer
tkinter-tix
streamlit_webrtc
PIL
numpy
cv2
threading
pandas
random
requests
json
aiortc-contrib-media
time
tkinter
turtle
seaborn
matplotlib-pyplot
google-cloud
io